+++ classes/warrior.lua
@@ -568,6 +568,38 @@ local function useSwordAndBoard()
     );
 end
 
+local function useIncite()
+    local inciteBuff = 86627;
+    local inciteTalent = 50685; -- Rank 1
+
+    SAO:CreateEffect(
+        "incite",
+        SAO.CATA,
+        inciteBuff,
+        "aura",
+        {
+            talent = inciteTalent,
+            button = heroicStrike,
+        }
+    );
+end
+
+local function useGlyphOfIncite()
+    local glyphOfInciteBuff = 122016;
+    local glyphOfInciteTalent = 122013;
+
+    SAO:CreateEffect(
+        "glyph_of_incite",
+        SAO.MOP,
+        glyphOfInciteBuff,
+        "aura",
+        {
+            talent = glyphOfInciteTalent,
+            buttons = { heroicStrike, cleave },
+        }
+    );
+end
+
 local function useRegicide()
     local regicideBuff = 1231436;
 
@@ -601,6 +633,10 @@ local function registerClass(self)
 
     -- Protection
     useSwordAndBoard();
+    useIncite();
+
+    -- Talents
+    useGlyphOfIncite();
 
     -- Items
     useRegicide();